webchecklist/requirements.txt
asgiref==3.5.2
colorama==0.4.5
Django==4.1.1
django-auto-logout==0.5.0
django-axes==5.39.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-formtools==2.4
django-ipware==4.0.2
django-otp==1.1.3
django-phonenumber-field==6.4.0
django-recaptcha==3.0.0
django-two-factor-auth==1.14.0
phonenumbers==8.12.56
pytz==2022.4
qrcode==7.3.1
sqlparse==0.4.3
tzdata==2022.4
